// src/app/shared/model/AppInsuranceObj.Model.ts
export class AppInsuranceObj {
    AppInsId: number;
    AppId: number;
    TotalCustMainPremiAmt: number;
    TotalCustAddPremiAmt: number;
    TotalCustFeeAmt: number;
    TotalCustDiscAmt: number;
    TotalCustPremiAmt: number;
    TotalInscoMainPremiAmt: number;
    TotalInscoAddPremiAmt: number;
    TotalInscoFeeAmt: number;
    TotalInscoDiscAmt: number;
    TotalInscoPremiAmt: number;
    TotalAddInsPremiAmt: number;
    TotalInsCptlzAmt: number;
    TotalPremiPaidByCustAmt: number;
    TotalInsIncomeAmt: number;
    RowVersion: string;

constructor() { 
    this.AppInsId = 0;
    this.AppId = 0;
    this.TotalCustMainPremiAmt = 0;
    this.TotalCustAddPremiAmt = 0;
    this.TotalCustFeeAmt = 0;
    this.TotalCustDiscAmt = 0;
    this.TotalCustPremiAmt = 0;
    this.TotalInscoMainPremiAmt = 0;
    this.TotalInscoAddPremiAmt = 0;
    this.TotalInscoFeeAmt = 0;
    this.TotalInscoDiscAmt = 0;
    this.TotalInscoPremiAmt = 0;
    this.TotalAddInsPremiAmt = 0;
    this.TotalInsCptlzAmt = 0;
    this.TotalPremiPaidByCustAmt = 0;
    this.TotalInsIncomeAmt = 0;
    this.RowVersion = "";
    }
}